Extracted etl logic into separate modules.
